Added in the grid streching subroutines
Here I built added in the grid streching subroutines - initial results look good. Need to add in tests to check the surface metrics.

Main bits that have changed: the domain building module now computes the grid streched domain and the relevant metrics (1st order and second order numerically).

Then the equation builder need to change to account for the metrics. This is done in the scale subroutine.